Finalized the clean tests for the calculator API

+++ Samples/Calculator/Calculator.API/InputModels/V2/RectanglePerimeterInputModel.cs
@@ -2,6 +2,6 @@
 
 public class RectanglePerimeterInputModel
 {
-    public required decimal Side1 { get; init; }
-    public required decimal Side2 { get; init; }
+    public required int Side1 { get; init; }
+    public required int Side2 { get; init; }
 }
